# contrato/openapi.yaml
openapi: 3.0.0
paths:
...
  /yo:
    get:
      description: "La primera llamada de todas las pantallas: de `rol` sale a qué
        panel entra, `perfilCompleto` dice si hay que pedirle datos antes, y
        `edicion` evita tener que pedir la configuración aparte."
      operationId: PerfilesController_yo
      parameters: []
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/YoDto"
        "401":
          description: Sin sesión, o la sesión venció
      summary: Quién está logueado, con qué rol y en qué estado está el concurso
      tags: &a2
        - Perfil
    put:
      description: Lo que se manda reemplaza a lo que había; lo que no se manda queda
        igual. Es la pantalla que ve primero quien se inscribió solo con el
        correo.
      operationId: PerfilesController_actualizarDatos
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DatosPersonalesDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/YoDto"
      summary: Completa o corrige los datos personales desde el panel
      tags: *a2
  /inscripcion:
    post:
      description: Solo `correo`, `aceptaBases` y `turnstileToken` son obligatorios.
        Habilita en el acto para entrar con Google. Un correo que ya estaba
        recibe exactamente la misma respuesta que uno nuevo, a propósito. Ver
        docs/11-formulario-para-el-front.md.
      operationId: PerfilesController_inscribir
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InscripcionDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaInscripcionDto"
        "403":
          description: "Turnstile no aprobó el envío: resetear el widget"
        "429":
          description: Demasiados envíos desde la misma IP
      summary: El formulario de inscripción de la landing
      tags: *a2
  /inscripcion/completar/{token}:
    get:
      description: El token viene en el correo de alerta. Deja de servir cuando la
        inscripción se completa.
      operationId: PerfilesController_paraCompletar
      parameters:
        - name: token
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DatosParaCompletarDto"
        "404":
          description: "El token no existe o ya se usó: mostrar el formulario vacío"
      summary: Los datos para precargar el formulario desde «Completar mi inscripción»
      tags: *a2
  /r/{token}:
    get:
      description: "Anota el clic y redirige al grupo. Es el enlace de todos los
        correos y de la pantalla de gracias: así se frena el recordatorio, y el
        grupo se puede cambiar sin reenviar nada. Es una navegación, no un
        fetch."
      operationId: PerfilesController_alGrupo
      parameters:
        - name: token
          required: true
          in: path
          schema:
            type: string
      responses:
        "302":
          description: Redirige al grupo de WhatsApp
      summary: El enlace al grupo de WhatsApp
      tags: *a2
  /admin/jurados:
    post:
      description: Se invita uno por uno. Le llega un correo con el enlace para entrar
        con Google; tiene que usar la cuenta de ESE correo.
      operationId: PerfilesController_invitarJurado
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvitarJuradoDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JuradoDto"
        "409":
          description: El correo es de un admin o de alguien inscripto como concursante
      summary: Invita a un jurado
      tags: *a2
  /auth/google:
    get:
      description: "Redirige a Google (302). No es una llamada de fetch: el navegador
        tiene que navegar acá. `retorno` es la ruta del front a la que volver
        después de entrar. `invitacion` (el token del correo de invitación) o
        `equipo` (el token del enlace del equipo) dejan entrar a una cuenta de
        Google que todavía no está inscripta."
      operationId: AuthController_ingresar
      parameters:
        - name: invitacion
          required: true
          in: query
          schema:
            type: string
        - name: equipo
          required: true
          in: query
          schema:
            type: string
      responses:
        "200":
          description: ""
      summary: Arranca el ingreso con Google
      tags: &a3
        - Autenticación
...
  /mi-equipo:
    get:
      description: Devuelve `null` si todavía no armó ni se sumó a ninguno.
      operationId: EquiposController_mio
      parameters: []
      responses:
        "200":
          description: El equipo, o null
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EquipoDto"
      summary: El equipo de quien pregunta
      tags: &a4
        - Equipos
    post:
      description: Se puede crear vacío y sumar gente después. Crea también la
        propuesta en borrador. Responde 403 si quien lo pide no completó sus
        datos (`perfilCompleto` en `/yo`).
      operationId: EquiposController_crear
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CrearEquipoDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EquipoDto"
      summary: Arma un equipo
      tags: *a4
  /mi-equipo/invitaciones:
    post:
      description: Solo el líder. A cada dirección le llega un correo con su enlace.
        Devuelve los enlaces igual, por si hace falta pasarlos por otro lado.
      operationId: EquiposController_invitar
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvitarDto"
      responses:
        "201":
          description: ""
      summary: Invita integrantes por correo
      tags: *a4
  /invitacion/{token}:
    get:
      description: La pantalla `/invitacion/{token}` del front. Después, el botón
        manda a `/auth/google?invitacion={token}&retorno=/invitacion/{token}` y,
        de vuelta, llama a `POST /invitacion/{token}/aceptar`.
      operationId: EquiposController_verInvitacion
      parameters:
        - name: token
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitacionDto"
        "404":
          description: No existe, o ya se aceptó
      summary: Qué equipo invita, para mostrarlo antes del login
      tags: *a4
  /invitacion/{token}/aceptar:
    post:
      description: "Sirve con cualquier cuenta de Google, no solo la del correo
        invitado: el token es lo que identifica la invitación. Registra la
        aceptación de las bases con fecha, la versión vigente y la IP."
      operationId: EquiposController_aceptarInvitacion
      parameters:
        - name: token
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EquipoDto"
        "409":
          description: Ya forma parte de otro equipo
      summary: Acepta la invitación por correo
      tags: *a4
...
  /equipos/sumarme/{token}:
    post:
      description: "El camino sin fricción: sirve con cualquier cuenta de Google, sin
        que ningún correo tenga que coincidir. Para entrar sin estar inscripto,
        el login va con `/auth/google?equipo={token}`. Registra la aceptación de
        las bases con fecha, la versión vigente y la IP."
      operationId: EquiposController_sumarme
      parameters:
        - name: token
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AceptarInvitacionDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EquipoDto"
      summary: Se suma a un equipo con el enlace compartido
      tags: *a4
...
info:
  title: Habisite Design Challenge 2026 — API
  description: >-
    API del concurso. Tres roles y tres superficies: la inscripción pública, el
    panel

    de concursantes y el panel de jurado.


    ## Sesión


    Se entra con Google. El navegador tiene que **navegar** a `GET /auth/google`

    (un enlace, no un fetch); de ahí en más la sesión viaja en una cookie
    `HttpOnly`

    que el navegador manda sola.


    Todas las llamadas tienen que ir con `credentials: "include"`, o la cookie
    no viaja.


    ## Nadie entra sin estar inscripto


    Los perfiles no nacen del login: nacen del formulario de inscripción o de
    una

    invitación. Una cuenta de Google que no figure recibe `403`.


    ## Forma de los errores


    ```json

    { "estado": 404, "mensaje": "No existe esa propuesta", "ruta": "/…", "hora":
    "…" }

    ```


    Con `detalles: string[]` cuando la validación rechaza campos, un renglón por
    campo.
  version: "1.0"
  contact: {}
...
servers: []
components:
  securitySchemes:
    cookie:
      type: apiKey
      in: cookie
      name: sesion
  schemas:
...
    EdicionResumenDto:
      type: object
      properties:
        estado:
          type: string
          enum:
            - inscripcion
            - entregas
            - preseleccion
            - final
            - cerrada
            - publicada
        cierreEntregas:
          type: object
          nullable: true
          example: 2026-10-15T23:59:59-03:00
        margenGraciaMinutos:
          type: number
          description: Minutos de gracia después del cierre
          example: 15
        entregasAbiertas:
          type: boolean
          description: Lo único que el front necesita mirar para decidir si muestra el
            botón de entregar. Ya tiene en cuenta el margen de gracia.
        maxBytes:
          type: number
          example: 31457280
          description: Tope del PDF en bytes
        maxIntegrantes:
          type: object
          nullable: true
          example: 5
        maxPaginas:
          type: object
          nullable: true
        zonaHoraria:
          type: string
          example: America/Argentina/Buenos_Aires
        terminosUrl:
          type: object
          nullable: true
          description: "Dónde se leen las bases. `null` mientras no estén publicadas:
            mostrar el texto sin enlace."
      required:
        - estado
        - cierreEntregas
        - margenGraciaMinutos
        - entregasAbiertas
        - maxBytes
        - maxIntegrantes
        - maxPaginas
        - zonaHoraria
        - terminosUrl
    YoDto:
      type: object
      properties:
        id:
          type: string
        correo:
          type: string
        nombre:
          type: string
        apellido:
          type: string
        telefono:
          type: object
          nullable: true
          example: "+5491123456789"
        tipoInstitucion:
          type: string
          nullable: true
          enum:
            - universidad
            - trabajo
            - independiente
        institucion:
          type: object
          nullable: true
        pais:
          type: object
          nullable: true
        rol:
          type: string
          enum:
            - participante
            - jurado
            - admin
          description: Decide a qué panel entra.
        perfilCompleto:
          type: boolean
          description: "Si es `false`, el panel tiene que pedirle los datos que faltan
            (`PUT /yo`) antes de dejarlo armar equipo o subir algo: el servidor
            lo rechaza con 403 igual. Siempre `true` para jurados y admins."
        edicion:
          $ref: "#/components/schemas/EdicionResumenDto"
      required:
        - id
        - correo
        - nombre
        - apellido
        - telefono
        - tipoInstitucion
        - institucion
        - pais
        - rol
        - perfilCompleto
        - edicion
    DatosPersonalesDto:
      type: object
      properties:
        nombre:
          type: string
          example: Ana
        apellido:
          type: string
          example: Duarte
        telefono:
          type: string
          description: "Formato internacional E.164: + código de país y número, sin
            espacios."
          example: "+5491123456789"
        tipoInstitucion:
          type: string
          enum: &a8
            - universidad
            - trabajo
            - independiente
        institucion:
          type: string
          description: "Universidad o lugar de trabajo: puede haber profesionales jóvenes."
          example: FADU-UBA
        pais:
          type: string
          example: AR
          description: Código ISO de dos letras. `ZZ` para «Otro».
    InscripcionDto:
      type: object
      properties:
        nombre:
          type: string
          example: Ana
        apellido:
          type: string
          example: Duarte
        telefono:
          type: string
          description: "Formato internacional E.164: + código de país y número, sin
            espacios."
          example: "+5491123456789"
        tipoInstitucion:
          type: string
          enum: *a8
        institucion:
          type: string
          description: "Universidad o lugar de trabajo: puede haber profesionales jóvenes."
          example: FADU-UBA
        pais:
          type: string
          example: AR
          description: Código ISO de dos letras. `ZZ` para «Otro».
        correo:
          type: string
          description: "El correo de Google con el que va a entrar. Tiene que ser
            exactamente ese: si después ingresa con otra cuenta, queda afuera."
          example: ana@gmail.com
        aceptaBases:
          type: boolean
          description: Tiene que ser `true`. La versión de las bases la registra el
            servidor.
          example: true
        turnstileToken:
          type: string
          description: El token que genera el widget de Cloudflare Turnstile.
        origen:
          type: string
          description: "De qué canal vino: lo lee el front del parámetro `?origen=` de la
            URL."
          example: linkedin
        tokenCompletar:
          type: string
          description: Solo si la landing se abrió desde el botón «Completar mi
            inscripción» del correo (`?completar=`). Completa esa inscripción en
            vez de crear otra.
      required:
        - correo
        - aceptaBases
        - turnstileToken
    RespuestaInscripcionDto:
      type: object
      properties:
        estado:
          type: string
          enum:
            - completa
            - incompleta
          description: "`completa`: redirigir al grupo a los 3 segundos. `incompleta`:
            mostrar el botón al grupo y avisar que le llega un correo para
            completar. Depende solo de lo que se mandó en este envío."
        whatsapp:
          type: string
          description: El enlace al grupo de WhatsApp. Pasa por la API (registra el clic)
            y de ahí redirige. Usar siempre este, nunca uno escrito en el front.
          example: https://api.challenge.habisite.com/r/Xk3f…
      required:
        - estado
        - whatsapp
    DatosParaCompletarDto:
      type: object
      properties:
        correo:
          type: string
        nombre:
          type: string
        apellido:
          type: string
        telefono:
          type: object
          nullable: true
        tipoInstitucion:
          type: string
          nullable: true
          enum: *a8
        institucion:
          type: object
          nullable: true
        pais:
          type: object
          nullable: true
      required:
        - correo
        - nombre
        - apellido
        - telefono
        - tipoInstitucion
        - institucion
        - pais
    InvitarJuradoDto:
      type: object
      properties:
        correo:
          type: string
          description: "El correo de su cuenta de Google: con ese y solo con ese va a
            poder entrar."
          example: jurado@gmail.com
        nombre:
          type: string
          example: Sol
        apellido:
          type: string
          example: Pulido
      required:
        - correo
    JuradoDto:
      type: object
      properties:
        id:
          type: string
        correo:
          type: string
        nombre:
          type: string
        apellido:
          type: string
        estado:
          type: string
          description: "`habilitado` hasta que entre por primera vez; después `activo`."
      required:
        - id
        - correo
        - nombre
        - apellido
        - estado
    MiembroDto:
      type: object
      properties:
        perfilId:
          type: string
        nombre:
          type: string
        apellido:
          type: string
        correo:
          type: string
        institucion:
          type: object
          nullable: true
        estado:
          type: string
          enum:
            - invitado
            - aceptado
          description: "`invitado` todavía no aceptó: figura en la lista pero no cuenta
            como autor."
        esLider:
          type: boolean
      required:
        - perfilId
        - nombre
        - apellido
        - correo
        - institucion
        - estado
        - esLider
    EquipoDto:
      type: object
      properties:
        id:
          type: string
        nombre:
          type: object
          nullable: true
        miembros:
          type: array
          items:
            $ref: "#/components/schemas/MiembroDto"
        enlaceInvitacion:
          type: object
          nullable: true
          description: El enlace para compartir. Solo lo ve el líder; a los demás les
            llega `null`.
        invitacionActiva:
          type: boolean
        soyLider:
          type: boolean
          description: Si quien pregunta es el líder del equipo.
        maxIntegrantes:
          type: object
          nullable: true
          description: Tope de integrantes, o `null` si no hay.
      required:
        - id
        - nombre
        - miembros
        - enlaceInvitacion
        - invitacionActiva
        - soyLider
        - maxIntegrantes
    CrearEquipoDto:
      type: object
      properties:
        nombre:
          type: string
          description: Se puede poner después.
          example: Estudio Norte
    InvitarDto:
      type: object
      properties:
        correos:
          description: Un correo por integrante. A cada uno le llega un enlace que sirve
            con cualquier cuenta de Google, no solo la de ese correo.
          example:
            - tomas@gmail.com
            - sol@gmail.com
          type: array
          items:
            type: string
      required:
        - correos
    InvitacionDto:
      type: object
      properties:
        equipo:
          type: object
          nullable: true
          example: Estudio Norte
        invitadoPor:
          type: object
          nullable: true
          example: Ana Duarte
          description: Quién armó el equipo. `null` si todavía no cargó su nombre.
      required:
        - equipo
        - invitadoPor
    AceptarInvitacionDto:
      type: object
      properties:
        terminosVersion:
          type: string
          deprecated: true
          description: "Ya no se usa: la versión de las bases la pone el servidor (la
            vigente en la edición). Se acepta para no romper llamadas viejas,
            pero se ignora."
...